Move all events to OmniSharp.Models.Events namespace
