+++ src/OmniSharp.Abstractions/Extensions/IEventEmitterExtensions.cs
@@ -1,4 +1,4 @@
-﻿using OmniSharp.Models;
+﻿using OmniSharp.Models.Events;
 
 namespace OmniSharp.Services
 {
