+++ src/OmniSharp.Abstractions/Models/Events/EventTypes.cs
@@ -1,4 +1,4 @@
-namespace OmniSharp.Models
+namespace OmniSharp.Models.Events
 {
     public static class EventTypes
     {
