+++ src/OmniSharp.Abstractions/Models/Events/PackageDependency.cs
@@ -1,4 +1,4 @@
-namespace OmniSharp.Models
+namespace OmniSharp.Models.Events
 {
     public class PackageDependency
     {
