+++ src/OmniSharp.DotNet/Cache/ProjectStatesCache.cs
@@ -6,7 +6,7 @@ using Microsoft.DotNet.ProjectModel;
 using Microsoft.Extensions.Logging;
 using NuGet.Frameworks;
 using OmniSharp.DotNet.Models;
-using OmniSharp.Models;
+using OmniSharp.Models.Events;
 using OmniSharp.Models.ProjectInformation;
 using OmniSharp.Services;
